penambahan fitur pencarian

+++ Pkl/admin/guru.php
@@ -1,6 +1,6 @@
 <?php
-    include '../koneksi.php';
-    include 'header.php';
+include '../koneksi.php';
+include 'header.php';
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -22,7 +22,21 @@
             <h4>Daftar Guru</h4>
         </div>
         <div class="panel-body">
-            <a href="tambah_guru.php" class="btn btn-primary">Tambah Guru</a>
+            <div class="mt-2">
+                <div class="row">
+                    <div class="col-md-2">
+                        <a href="tambah_guru.php" class="btn btn-primary">Tambah Guru</a>
+                    </div>
+                    <div class="col-md-8">
+                        <form action="guru.php" method="get">
+                            <div class="input-group">
+                                <input type="text" name="cari" class="form-control ml-3" placeholder="Cari berdasarkan nama guru..." aria-label="Search" aria-describedby="search-icon" value="<?php if(isset($_GET['cari'])){ echo $_GET['cari'];} ?>">
+                                <button class="btn btn-primary" type="submit" id="search-icon"><i class="bi bi-search"></i></button>
+                            </div>
+                        </form>         
+                    </div>
+                </div>
+            </div>
             <br><br><br>
             <table class="table table-bordered table-striped">
                 <tr>
@@ -39,27 +53,48 @@
                     <th class="text-center" width="10%">Opsi</th>
                 </tr>
                 <?php
-                    $data = mysqli_query($koneksi, "SELECT * FROM guru ORDER BY nip ASC");
-                    $no = 1;
-                    while ($d = mysqli_fetch_array($data)) {
-                        ?>
-                        <tr>
-                            <td><?php echo $no++ ?></td>
-                            <td><?php echo $d['nip']; ?></td>
-                            <td><?php echo $d['nama']; ?></td>
-                            <td><?php echo $d['alamat']; ?></td>
-                            <td><?php echo $d['telp']; ?></td>
-                            <td><?php echo $d['password']; ?></td>
-                            <td><?php echo $d['tahun']; ?></td>
-                            <td><?php echo $d['id_jurusan']; ?></td>
-                            <td><?php echo $d['jabatan']; ?></td>
-                            <td><?php echo $d['pangkat']; ?></td>
-                            <td>
-                                <a href="edit_guru.php?nip=<?php echo $d['nip']; ?>" class="btn btn-sm btn-info">Edit</a>
-                                <a href="hapus_guru.php?nip=<?php echo $d['nip']; ?>" class="btn btn-sm btn-danger">Hapus</a>
-                            </td>
-                        </tr>
-                        <?php
+                    $query = "SELECT * FROM guru";
+                    if(isset($_GET['cari'])){
+                        $cari = $_GET['cari'];
+                        // Filter data guru berdasarkan nama atau NIP yang mengandung nilai pencarian
+                        $query .= " WHERE nama LIKE '%$cari%' OR nip LIKE '%$cari%'";
+                    }
+                    
+                    // Urutkan data guru berdasarkan NIP secara ascending
+                    $query .= " ORDER BY nip ASC";
+                    
+                    $result = mysqli_query($koneksi, $query);
+                    
+                    // Periksa apakah query dieksekusi dengan sukses
+                    if(!$result) {
+                        die("Query Error: " . mysqli_errno($koneksi) . " - " . mysqli_error($koneksi));
+                    }
+                    
+                    // Periksa apakah ada data yang ditemukan
+                    if(mysqli_num_rows($result) == 0) {
+                        echo '<tr><td colspan="11" class="text-center">Data tidak ditemukan</td></tr>';
+                    } else {
+                        $no = 1;
+                        while ($d = mysqli_fetch_array($result)) {
+                            ?>
+                            <tr>
+                                <td><?php echo $no++ ?></td>
+                                <td><?php echo $d['nip']; ?></td>
+                                <td><?php echo $d['nama']; ?></td>
+                                <td><?php echo $d['alamat']; ?></td>
+                                <td><?php echo $d['telp']; ?></td>
+                                <td><?php echo $d['password']; ?></td>
+                                <td><?php echo $d['tahun']; ?></td>
+                                <td><?php echo $d['id_jurusan']; ?></td>
+                                <td><?php echo $d['jabatan']; ?></td>
+                                <td><?php echo $d['pangkat']; ?></td>
+                                <td>
+                                    <a href="edit_guru.php?nip=<?php echo $d['nip']; ?>" class="btn btn-sm btn-info">Edit</a>
+                                    <a href="hapus_guru.php?nip=<?php echo $d['nip']; ?>" class="btn btn-sm btn-danger">Hapus</a>
+                                </td>
+                            </tr>
+                            <?php
+                        }
                     }
                 ?>
             </table>
